fixed some login bugs
