+++ src/hooks/useUser.tsx
@@ -1,9 +1,35 @@
-import { createContext, useContext } from "react";
-import type { User } from "firebase/auth";
+import { onAuthStateChanged, type User } from "firebase/auth";
+import { createContext, useContext, useState, useEffect } from "react";
+import { auth } from "../../firebaseConfig";
 
 type UserContextType = {
-    user: User | null;
+  user: User | null;
+  loading: boolean;
 };
 
-export const UserContext = createContext<UserContextType>({ user: null });
+const UserContext = createContext<UserContextType>({
+  user: null,
+  loading: true,
+});
+
+export const UserProvider = ({ children }: { children: React.ReactNode }) => {
+  const [user, setUser] = useState<User | null>(null);
+  const [loading, setLoading] = useState(true);
+
+  useEffect(() => {
+    const unsubscribe = onAuthStateChanged(auth, (firebaseUser) => {
+      setUser(firebaseUser);
+      setLoading(false);
+    });
+
+    return () => unsubscribe();
+  }, []);
+
+  return (
+    <UserContext.Provider value={{ user, loading }}>
+      {children}
+    </UserContext.Provider>
+  );
+};
+
 export const useUser = () => useContext(UserContext);
