+++ src/pages/MultiPlayer.tsx
@@ -9,11 +9,11 @@ const MultiPlayer: React.FC = () => {
 
   const navigate = useNavigate();
 
-  const { user } = useUser()
-  
+  const { user, loading } = useUser()
+    
   useEffect(() => {
-      if(!user) navigate("/login");
-  })
+      if(!user && !loading) navigate("/login");
+  }, [])
 
   const handleCreateRoom = () => {
     const roomId = Math.floor(Math.random() * 100000) + 100000;
