--- src/utils/userProvider.tsx
@@ -1,22 +1,0 @@
-import { useState, useEffect } from "react";
-import { onAuthStateChanged } from "firebase/auth";
-import type { User } from "firebase/auth";
-import { auth } from "../../firebaseConfig";
-import { UserContext } from "../hooks/useUser";
-
-export const UserProvider = ({ children }: { children: React.ReactNode }) => {
-    const [user, setUser] = useState<User | null>(null);
-
-    useEffect(() => {
-        const unsub = onAuthStateChanged(auth, (firebaseUser) => {
-            setUser(firebaseUser);
-        });
-        return () => unsub();
-    }, []);
-
-    return (
-        <UserContext.Provider value={{ user }}>
-            {children}
-        </UserContext.Provider>
-    );
-}
